+++ src/Contracts/BodyPayload/BodyPayloadContract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EasyHTTP\Contracts\Contracts\BodyPayload;
+
+interface BodyPayloadContract
+{
+    public function getContents();
+}