+++ src/Contracts/BodyPayload/IterableBodyPayloadContract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EasyHTTP\Contracts\Contracts\BodyPayload;
+
+interface IterableBodyPayloadContract extends BodyPayloadContract
+{
+    public function getContents(): iterable;
+}