+++ src/Contracts/BodyPayload/ResourceBodyPayloadContract.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace EasyHTTP\Contracts\Contracts\BodyPayload;
+
+interface ResourceBodyPayloadContract extends BodyPayloadContract
+{
+    /**
+     * @return resource
+     */
+    public function getContents();
+}